+++ _pages/biomed.html
@@ -9,11 +9,11 @@ permalink: /biomed/
     <title>BioMed Journal Club</title>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script type="module" crossorigin src="{{ '/assets/react/biomed-Ds939vcm.js' | relative_url }}"></script>
-    <link rel="modulepreload" crossorigin href="{{ '/assets/react/fluent-carousel-EziJ4bD5.js' | relative_url }}">
-    <link rel="modulepreload" crossorigin href="{{ '/assets/react/index-ID_pyUxO.js' | relative_url }}">
-    <link rel="modulepreload" crossorigin href="{{ '/assets/react/Card-CiKJtrWm.js' | relative_url }}">
-    <link rel="modulepreload" crossorigin href="{{ '/assets/react/linkContext-CcisxyYK.js' | relative_url }}">
+    <script type="module" crossorigin src="{{ '/assets/react/biomed-f7tzDJmv.js' | relative_url }}"></script>
+    <link rel="modulepreload" crossorigin href="{{ '/assets/react/fluent-carousel-D7IcKjT-.js' | relative_url }}">
+    <link rel="modulepreload" crossorigin href="{{ '/assets/react/index-CvylHR8S.js' | relative_url }}">
+    <link rel="modulepreload" crossorigin href="{{ '/assets/react/Card-Br9fYQyl.js' | relative_url }}">
+    <link rel="modulepreload" crossorigin href="{{ '/assets/react/linkContext-BM-iLui6.js' | relative_url }}">
     <link rel="stylesheet" crossorigin href="{{ '/assets/react/index-DQ3P1g1z.css' | relative_url }}">
   </head>
   <body>
