+++ react/src/team/index.tsx
@@ -1,310 +1,7 @@
-import { StrictMode, type ReactNode } from 'react'
+import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import {
-  FluentProvider,
-  webLightTheme,
-  Persona,
-  Divider,
-  Image,
-  Link,
-  makeStyles,
-  tokens,
-  type AvatarSize,
-} from '@fluentui/react-components'
-import {
-  Carousel,
-  CarouselCard,
-  CarouselNav,
-  CarouselNavContainer,
-  CarouselNavImageButton,
-  CarouselSlider,
-  CarouselViewport,
-} from '@fluentui/react-carousel'
 import '../index.css'
-import people from './people.json'
-import photo from './photo.json'
-
-const useStyles = makeStyles({
-  app: {
-    padding: '2rem 1.5rem',
-    maxWidth: '960px',
-    margin: '0 auto',
-    display: 'grid',
-    rowGap: '2.5rem',
-  },
-  section: {
-    display: 'grid',
-    rowGap: '1.5rem',
-  },
-  list: {
-    display: 'grid',
-    rowGap: '1rem',
-  },
-})
-
-const ASSETS_BASE = __DEV__ ? '/assets/' : '/assets/react/'
-
-const usePersonaCardStyles = makeStyles({
-  secondaryBlock: {
-    display: 'grid',
-    rowGap: '0.25rem',
-    justifyItems: 'flex-start',
-  },
-  metaLine: {
-    color: tokens.colorNeutralForeground2,
-    '& strong': {
-      color: tokens.colorNeutralForeground1,
-    },
-    '& ul': {
-      margin: '0.25rem 0 0 1.25rem',
-    },
-  },
-  linkRow: {
-    display: 'flex',
-    flexWrap: 'wrap',
-    columnGap: '0.75rem',
-    rowGap: '0.25rem',
-    justifyContent: 'flex-start',
-  },
-})
-
-const PersonaCard = ({
-  person,
-  size = 96,
-  lines = [],
-}: {
-  person: Person
-  size?: AvatarSize,
-  lines?: Array<ReactNode | undefined>
-}) => {
-  const styles = usePersonaCardStyles()
-  const email = person.email ?? person.contact?.email
-  const phone = person.contact?.phone
-  const linkItems = [
-    email ? (
-      <Link key="email" href={`mailto:${email}`}>
-        {email}
-      </Link>
-    ) : null,
-    phone ? (
-      <Link key="phone" href={`tel:${phone.replace(/[^\d+]/g, '')}`}>
-        {phone}
-      </Link>
-    ) : null,
-    person.linkedin ? (
-      <Link key="linkedin" href={person.linkedin} target="_blank" rel="noreferrer">
-        LinkedIn
-      </Link>
-    ) : null,
-    person.google_scholar ? (
-      <Link key="google" href={person.google_scholar} target="_blank" rel="noreferrer">
-        Google Scholar
-      </Link>
-    ) : null,
-    person.research_gate ? (
-      <Link key="researchgate" href={person.research_gate} target="_blank" rel="noreferrer">
-        ResearchGate
-      </Link>
-    ) : null,
-    person.cv ? (
-      <Link key="cv" href={person.cv} target="_blank" rel="noreferrer">
-        CV
-      </Link>
-    ) : null,
-  ].filter(i => !!i)
-  const infoLines = lines
-    .map((line) => {
-      if (line == null) {
-        return undefined
-      }
-      if (typeof line === 'string') {
-        const trimmed = line.trim()
-        return trimmed.length > 0 ? trimmed : undefined
-      }
-      return line
-    })
-    .filter(line => !!line)
-  const hasContent = infoLines.length > 0 || linkItems.length > 0
-
-  return (
-    <Persona
-      name={person.name}
-      size="huge"
-      avatar={person.avatar_filename ? {
-        size,
-        image: { src: `${ASSETS_BASE}people/${person.avatar_filename}` },
-        color: 'colorful',
-      } : undefined}
-      secondaryText={
-        hasContent
-          ? <div className={styles.secondaryBlock}>
-            {infoLines.map((line, index) => (
-              <div key={index} className={styles.metaLine}>
-                {line}
-              </div>
-            ))}
-            {linkItems.length > 0 && <div className={styles.linkRow}>{linkItems}</div>}
-          </div>
-          : undefined
-      }
-    />
-  )
-}
-
-const FacultySection = () => {
-  const styles = useStyles()
-  return (
-    <section className={styles.section}>
-      <h1>Faculty</h1>
-      <div className={styles.list}>
-        {people.faculty.map((person) => (
-          <PersonaCard
-            key={person.name}
-            person={person}
-            size={128}
-            lines={[
-              <b>{person.title}</b>,
-              <b>{person.department}</b>,
-              <b>Education:</b>,
-              <ul>
-                {person.education?.map((edu) => (
-                  <li key={edu}>{edu}</li>
-                ))}
-              </ul>
-            ]}
-          />
-        ))}
-      </div>
-    </section>
-  )
-}
-
-const StudentsSection = () => {
-  const styles = useStyles()
-  const subsections: Array<[string, Person[]]> = [
-    ['PhD Students', people.current_students.phd_students],
-    ['Undergraduate Students', people.current_students.undergraduate_students],
-    ['High School Students', people.current_students.high_school_students],
-  ]
-
-  return (
-    <section className={styles.section}>
-      <h1>Current Students</h1>
-      {subsections.map(([label, entries]) => (
-        <div key={label} className={styles.section}>
-          <h2>{label}</h2>
-          <div className={styles.list}>
-            {entries.map((person) => (
-              <PersonaCard
-                key={person.name}
-                person={person}
-                lines={[
-                  person.study ?? person.school,
-                  person.research_topics,
-                  person.awards?.length
-                    ? (
-                      <>
-                        <strong>Awards:</strong>
-                        <ul>
-                          {person.awards.map((award) => (
-                            <li key={award}>{award}</li>
-                          ))}
-                        </ul>
-                      </>
-                    )
-                    : undefined,
-                ]}
-              />
-            ))}
-          </div>
-        </div>
-      ))}
-    </section>
-  )
-}
-
-const AlumniSection = () => {
-  const styles = useStyles()
-  const subsections: Array<[string, Person[]]> = [
-    ['M.S. Alumni', people.alumni.ms_students],
-    ['Undergraduate Alumni', people.alumni.undergraduate_students],
-    ['High School Alumni', people.alumni.high_school_students],
-  ]
-
-  return (
-    <section className={styles.section}>
-      <h1>Alumni</h1>
-      {subsections.map(([label, entries]) => (
-        <div key={label} className={styles.section}>
-          <h2>{label}</h2>
-          <div className={styles.list}>
-            {entries.map((person) => (
-              <PersonaCard
-                key={`${person.name}-${person.duration ?? 'alumni'}`}
-                person={person}
-                lines={[person.department ?? person.school, person.duration]}
-              />
-            ))}
-          </div>
-        </div>
-      ))}
-    </section>
-  )
-}
-
-const PhotoSection = () => {
-  const styles = useStyles()
-  return (
-    <section className={styles.section}>
-      <h1>Photos</h1>
-      <Carousel>
-        <CarouselViewport>
-          <CarouselSlider>
-            {photo.map((img, index) => (
-              <CarouselCard key={index}>
-                <Image
-                  src={`${ASSETS_BASE}photo/${img.filename}`}
-                  alt={img.title ?? `Photo ${index + 1}`}
-                  width={600}
-                  height={400}
-                  fit="contain"
-                />
-                {img.title && <p>{img.title}</p>}
-              </CarouselCard>
-            ))}
-          </CarouselSlider>
-        </CarouselViewport>
-        <CarouselNavContainer>
-          <CarouselNav>
-            {index => (
-              <CarouselNavImageButton
-                key={index}
-                image={{ src: `${ASSETS_BASE}photo/${photo[index].filename}` }}
-              />
-            )}
-          </CarouselNav>
-        </CarouselNavContainer>
-      </Carousel>
-    </section>
-  )
-}
-
-const App = () => {
-  const styles = useStyles()
-  return (
-    <FluentProvider theme={webLightTheme}>
-      <main className={styles.app}>
-        <FacultySection />
-        <Divider />
-        <StudentsSection />
-        <Divider />
-        <AlumniSection />
-        <Divider />
-        <PhotoSection />
-      </main>
-    </FluentProvider>
-  )
-}
+import { App } from './app'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
